Make FB.Event usable as an FB.EventProvider mixin
Summary:

The existing implementation used an absolute reference to the subscriber
object. This allowed the function to be used standalone without being
explicitly attached to FB.Event. Meaning this would work:

  var subscribe = FB.Event.subscribe;
  subscribe('auth.login', function() {});

While this was intentional, its not something you see often and certainly not
something the library itself was using.

In order to share the Event system between the Global Named Events
('auth.login', 'xfbml.rendered') and for instance objects, we change it
to use the subscriber object from the current scope. This is basically
the change contained here.

The result, is that we can do something like:

  var MyClass = function() {};
  FB.copy(MyClass.prototype, FB.EventProvider);

  var anInstance = new MyClass();
  anInstance.subscribe('dataReady', function(r) { console.log('dataReady', r) });
  anInstance.fire('dataReady', 42);

With the new logic, global event subscribers are stored in
FB.Event._subscribersMap, where as the subscribers for anInstance above are stored
in anInstance._subscribersMap.

Task ID: #

Blame Rev:

Reviewers: wzhu, lshepard, ahimel, mvernal

CC: platform-diffs@lists.facebook.com

Test Plan: Ran the unit tests. All passes. Added some tests to ensure the
functionality works as expected when used with in a mixin style.

Revert Plan:

Tags:

Database Impact:

Memcache Impact:

Other Notes:

EImportant:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

DiffCamp Revision: 71331