Include json2.js in repository, add missing fb.event @requires
Summary:

Including json2.js in the repository for convenience. It's unmodified except
for the Haste docblock. The fb.event @requires were also missing.

Reviewed By: ahimel

CC: platform-diffs@lists.facebook.com

Test Plan:
The Loader in not in yet, but loaded it using the new assets.php loader here:
http://www.naitik.dev575.snc1.facebook.com/assets.php/en_US/init.js

It gives a working copy of the library.

DiffCamp Revision: 72077