Updating cookie format to be quoted to follow the RFC
Summary:
We were including the '=' sign in our cookie. This is not valid according to
the RFC unless the value is quoted. This breaking change adds those quotes to
our cookie value. This was reported on Github with relation to Apache Tomcat
which follows the RFC:
http://github.com/facebook/connect-js/issues/#issue/2/comment/77244

This is the first breaking change since the initial alpha release. I've created
a changelog.md file and linked it from the readme.

Reviewed By: lshepard

CC:platform-diffs@lists.facebook.com

Test Plan: Ran tests.

DiffCamp Revision: 75467