Add unit tests for XFBML parser
Summary:
Wrote unit tests for type system and XFBML framework in Joey. This diff does NOT include any tests for the actual tags. We are at 100% coverage for element.js, type.js, waitable.js, and close enough for xfbml.js - except for some statements which are error conditions and specific browser tests.

Also, in doing this I changed a little bit of the way we load tags. Rather than having the component be specified explicitly (which I think is redundant), just change the component name to always be a function of the tag name. That way it can be dynamically determined.

I also want as a design goal that the system should be able to work without dynamic loading - it's exclusively an optional feature, especially since it involves a server-side component. So now, instead of checking if the component has been loaded (which means something has to call FB.Component.scriptLoaded), instead we just check to see if the class has been defined - and if so, just go with that.

Reviewed By: wzhu

Test Plan:
Ran all the non-UI unit tests in FF3, Safari, and IE8. Saw code coverage and they all passed. (Except for some of the data ones, but that's a server bug that Ray is working on)

Revert: OK

DiffCamp Revision: 78925