XFBML unit tests about halfway
Summary:
After much experimentation I settled on a regex approach for the XFBML unit tests.

This diff gets us near 100% coverage for login-button and profile-pic, and most coverage for fb:name (except for all the pronoun combos), and not much for fb:share.

Also ... it doesn't work in all browsers. In FF3 they pass but in Safari and IE7, not always. In Safari they don't even render.

Reviewed By: wzhu

Test Plan:
Ran the unit tests in FF3, Safari, and IE7.

For some reason the regexes in IE7 don't pass all the time, so I get some failures, and in Safari the XFBML tags don't render.

http://www.lshepard-connect-js.devrs006.facebook.com/fbconnect/tests/index.html?fb:

Revert: OK

DiffCamp Revision: 80333