Ensure Cookie is set before auth events are fired
Summary:
This was reported by Justin Shaffer:
  http://github.com/facebook/connect-js/issues/#issue/40

Since the Cookie system relies on the auth events, it was possible for
user specified event handlers to be invoked before the internal cookie
handler was invoked. This caused a race condition where a user handler
may fire a XHR request before the cookie has been set, resulting in the
server not getting a cookie.

The other approach to fixing this was to add a way to register internal
handlers to be executed before user specified ones. This doesn't solve
the problem in this case, because there are different auth events. For
example, the "auth.login" event is fired before the "auth.sessionChange"
event. Since the cookie subscribes to the "auth.sessionChange" event,
this would mean, that even with a higher priority, a user subscribed
"auth.login" handler would get invoked before a high priority
"auth.sessionChange" event. So I opted to not rely on the event at all
and instead create a weak dependency to the cookie system.

Task ID: #

Blame Rev:

Reviewed By: brent

Other Commenters: ahimel

CC:platform-diffs@lists.facebook.com

Test Plan:Ran the existing test suite and updated the cookie test.

Revert Plan:

Tags:

Database Impact:

Memcache Impact:

Other Notes:

EImportant:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

DiffCamp Revision: 83793