callback hook for easier async loading support
Summary:
If a global named fbAsyncInit is defined, it will be invoked when the
library is loaded. This is useful when we want to load the library
asynchronously.

Task ID: #

Blame Rev:

Reviewed By: wzhu

Other Commenters: mkinsey

CC: mkinsey@facebook.com, platform-diffs@lists.facebook.com

Test Plan: Created a test page:
http://connect.daaku.org/new.html

I dont have a test case because of the complexity involved in testing
this with our current test setup.

Revert Plan:

Tags:

Database Impact:

Memcache Impact:

Other Notes:

EImportant:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

DiffCamp Revision: 94349