Graph Server Support for Connect JS
Summary:
This implements graph server support, but NOT OAuth support. There's a
counterpart diff which has the necessary server side changes. This diff
is intended to be backward compatible with the library save one change
-- the session JSON blob now includes an oauth_access_token. One
additional, hopefully transparent change is also the use of
oauth_access_token instead of signed calls to restserver.php (thanks
abanker!).

The API signature is basically:

  FB.api(path, method, params, cb)

Everything but the path is optional. So all of these are valid:

  FB.api('/me') // throw away the response
  FB.api('/me', function(r) { console.log(r) })
  FB.api('/me', { fields: 'email' }) // throw away response
  FB.api('/me', { fields: 'email' }, function(r) { console.log(r) })
  FB.api('/me/feed', 'post', { body: 'hi there' }, function(r) { console.log(r) })
  FB.api('/12345678', 'delete', function(r) { console.log(r) })

More readable docs here:
http://developers.naitik.dev575.snc1.facebook.com/docs/?u=facebook.joey.FB.api

Test Plan:
A bunch of examples to play with here:
http://fbrell.com/secret/graph-me?server=sb

Added a bunch of tests and ran the test suite.

DiffCamp Revision: 99715
Reviewed By: lshepard
Commenters: abanker
CC: platform-diffs@lists, lshepard, brent, naitik, abanker, mattwkelly, selekman
Revert Plan:
OK