i18n support for new connect
Summary:
Ported the i18n machinery from the old connect into the new one with a
bunch of namespacing and renaming. Server side stuff was already setup.

Since we make this library available via github, I have introduced a
string table like the main site usually has. This is currently en_US
only, and is a short term placeholder. This is not used when we serve
the library via Haste.

Iframe widget's currently do not get the locale automatically passed in,
I will be cleaning up how widgets provide their URLs and make that
automatically happen in a separate diff.

Test Plan:
Ran
./scripts/cachearchiver/regenerate_archives
./scripts/static_resources/update_database --test

[this failed with a bunch of errors, and took over 1 hour 40 mins! but
it seems to have done what i needed it to do...]

Load the french version of the library:
http://fbrell.com/xfbml/fb:share-button?locale=fr_FR&server=sb

- notice the <fb:share-button> is localized
- make sure you are not in the 'connected' state
- click on the share button, window popup is in the fr_FR locale
- click the login button, notice the login popup is in the fr_FR locale
- get into the 'connected' state
- click on the share button, notice the "Loading..." text is in the
  fr_FR locale
- notice the iframe dialog is in the fr_FR locale

Did the same for en_US.

DiffCamp Revision: 96407
Reviewed By: jleszcze
CC: platform-diffs@lists, lshepard, brent, jleszcze, naitik, selekman
Revert Plan:
OK