Pass correct parameters through new Connect library.
Summary:
Title says it all.  We changed 'external_url' and
'is_canonical_url' to 'href' and 'is_permalink' a while ago,
but it was never synced with the new Connect library.  This
does that.

Test Plan:
Loaded a new page in my sandbox using the new Connect library,
and it worked just swell.

DiffCamp Revision: 96566
Reviewed By: naitik
CC: naitik
Revert Plan:
OK