Fix race condition in auth enabled iframe widgets
Summary:
If a widget needs to refresh on auth change, it needs to know the auth
status when it was first rendered to decide if it has changed when an
event fires. The bug was that we were not relying on the actual current
status, but just the default 'unknown' value on load. This fixes it by
waiting for the initial login status.

Test Plan:
Loading this page in a notConnected state will fail to
render the widget:
http://fbrell.com/xfbml/fb:comments
fixed after this diff:
http://fbrell.com/xfbml/fb:comments?server=sb

DiffCamp Revision: 102495
Reviewed By: jcain
CC: platform-diffs@lists, jcain, mkinsey
Revert Plan:
OK