Drop fairly high z-indices on the comment widget iframe.
Summary:
When two widgets are stacked one above the other, we're
seeing the comment widget extend down into the other like widget,
and even though it's clear the comment widget should be at a higher
z-index that the other widget, it isn't.  This forces the z-index to
be a reasonably high number, and I increment them so that as they
open they must layer over previously opened ones (unlikely, but
technically possible with the smaller widgets.)

Test Plan:
Loaded www.jerrycain.com/G2213.html in my sandbox as a
test user, and everything works perfectly well now.  Tested in
all Mac browsers, IE8, and it worked fine.  Can't get it to work
in IE7.  Looking for suggestions on that one.

DiffCamp Revision: 103794
Reviewed By: naitik
CC: ahaugen, naitik
Revert Plan:
OK