Make sizes consistent between the PHP and Connect JS lib.
Summary:
We allowed the new Connect library to fall out of
sync with the PHP.  We need the defaults to be replicated in
both places.

Test Plan:
Loaded G200.html in my sandbox, changing
the width parameter to include various sizes.  Loaded
in Chrome and Safari (and Firefox) to confirm that
there are no longer any scrollbars.

DiffCamp Revision: 104003
Reviewed By: naitik
CC: naitik
Revert Plan:
OK