Fix compresed verison of FB.Data.waitOn
Summary:
There is a bug in jsxmin where it renames a variable to "c" even though
there is a "c" already defined in this scope. This is a work around for
that error until jsxmin can be fixed.

Test Plan:
I tested this in my sandbox by creating two compressed verisons of core.js
with jsxmin. One with the variable named "count" and one with it named "c".
Then I tested this on my test page @ http://alpjor.com/test2.html

DiffCamp Revision: 104036
Reviewed By: naitik
CC: mvernal, ahaugen
Revert Plan:
OK