xfbml for fb:like with button_count layout
Summary:
fb:like should display a number instead of text when used with the
parameter layout set to "button_count".
Faces should be disabled.
This needs to go to partners on 04/14.

Test Plan:
Go to fbrell, set the content as :
<fb:like href="cnn.com" layout="button_count"></fb:like>
<br/>
<fb:like href="cnn.com" layout="button_count"
colorscheme="dark"></fb:like>
<br/>
and interact with the button.

Note : there's currently a bug with the comment box iframe that capture
events even if it is not visible, but it is not related to this diff.
Still, if you try to like then unlike something, the invisible iframe
won't let you do it.

DiffCamp Revision: 104868
Reviewed By: naitik
Commenters: jleszcze
CC: ngao, ahaugen, cat, jleszcze, naitik
Tasks:
#191708

Revert Plan:
OK