IE bug when risizing window
Summary:
probably linked to this :
http://friendlybit.com/css/ie6-resize-bug/

if the containing window body has a fixed width, with margin auto,
and the iframe is contained within a floating div itself within a div
of width 50% of the body, then IE becomes crazy with the content of
our iframe on window resize.

Test Plan:
body {
  margin: 0 auto;
  width: 760px; /* Any fixed or fluid width */
  }

and in the body :
<body>
<div style="width:50%">
 <div style="float:right">
  <fb:recommendations site="cnn.com"></fb:recommendations>
 </div>
</div>
...

on IE load the page and resize the browser, you will see the content of
the iframe
being mis-aligned without the fix.

DiffCamp Revision: 106068
Reviewed By: jcain
CC: ngao, jcain
Tasks:
#194616

Revert Plan:
OK