xfbml side of like iframe showing/hiding comment
Summary:
like frame needs to notify the parent frame to apply/remove display:none
on the node containing the comment frame when appropriate.

Test Plan:
in fbrell, put
<fb:like href="http://www.jerrycain.com/C102.html"
is_permalink="true" action="recommend"></fb:like>
click on the like button, mouse in and out of the widget,
when the comment box is not visible, you should be able to
mouse over your pic in the facepile and see the tooltip.

DiffCamp Revision: 106479
Reviewed By: naitik
CC: ngao, naitik
Tasks:
#194541

Revert Plan:
OK