Connect Bar JS fix for position support in javascript libraries
Summary:
Yelp reported that when they tried to use the connect bar with their site they
found that absolutely positioned element were offset by the height of the
connect bar. This was because we were modifying the margin on the HTML tag as a
hack to get the body background image to move along with the body. In order to
fix this we needed to be able to animate the backgroundPosition css property and
that required a rewrite of the animation class.

I also found an instantce where an api callback was null and was firing an
error that I fixed.

I also fixed the css for the connect bar in ie6 that was busted from an earlier
commit.

Test Plan:
Tested in my sandbox:
http://alpjor.com/test2.html

Tested in all browsers
Tested the animation library for speed (much faster now)

DiffCamp Revision: 107015
Reviewed By: naitik
CC: mvernal, isaac, ahaugen, lshepard, brent, naitik, mattwkelly, alpjor,
selekman
Tasks:
#193035
#184421

Revert Plan:
OK