Add api-read support to joey.
Summary:
We only list a small number of calls in a micro-optimization to keep
transmission size down (though I didn't actually measure this).

Test Plan:
test page that makes an api call in the whitelist and not in the
whitelist; still works, and see that it hits the right endpoint in each
case.  UTs, but they only touch this obliquely (i.e., you can see that they
now hit api-read for the non-graph endpoints).

DiffCamp Revision: 107997
Reviewed By: naitik
Commenters: brent
CC: arudolph, lshepard, brent, naitik, mattwkelly, selekman
Revert Plan:
Ok