Custom XD Fragment and cleanup
Summary:
CNN & Amazon are asking for a custom XD Fragment URL. Generally, there
is one case where the Fragment transport fails -- when there is
background audio/video on the page, loading it in the iframe can cause
it to be play over there as well resulting in double audio streams
(video part would be hidden).

I've given postMessage the highest priority, but a custom channelUrl
will be prefered over Flash, where as without a custom channelUrl, we'll
still prefer Flash to reusing the current page.

I've also explicitly not documented the param in FB.init() for now, as I
think it will be confusing.

The channel file itself just needs to contain this line:

  <script src="http://connect.facebook.net/en_US/all.js"></script>

We could also suggest this URL instead, which results in a much smaller
JS file, but I'm not going to because the one above should already be
cached:

  http://connect.facebook.net/en_US/xd.js

Test Plan:
Working on adding tests..

DiffCamp Revision: 112249
Reviewed By: wzhu
CC: platform-diffs@lists, lshepard, brent, mattwkelly, wzhu, selekman
Tasks:

Revert Plan:
OK