auto refresh session every 20 minutes
Summary:
adds logic to setup a timer to ping login_status every 20 minutes. we
only do this if getLoginStatus() was invoked at least once. this ensures
that applications that do not want to fetch sessions in JS will not
get this timer. if the session will not expire, we do not setup a timer
either.

Test Plan:
For testing, I reduced the timeout from 20 minutes to 30 seconds. Then
tested that the timer is setup on page load, if a session is found, and
on login, if a session is returned.

DiffCamp Revision: 120550
Reviewed By: wzhu
CC: platform-diffs@lists, lshepard, brent, mattwkelly, wzhu, selekman
Revert Plan:
OK