connect js component of render-in-iframe for fb:login-button
Summary:
This allows the developer to pass render-in-iframe="true" to fb:login-button,
to force the widget to render in an iframe.

Test Plan:
in http://fbrell.com/?server=www.yariv.dev238, entered <fb:login-button
render-in-iframe="true"></fb:login>, and verified the button was rendered in an
iframe with no faces. Also checked some of the existing configs of fb:login and
fb:facepile to make sure they didn't break. (Btw, we really need to unit test
this stuff.)

DiffCamp Revision: 120838
Reviewed By: brent
CC: ahaugen, brent, yariv
Tasks:
#204567: create an iframe param for <fb:login-button>

Revert Plan:
OK